DT-835 Convert court controller tests to unit tests
